move getter to a separate file
