+++ extensions/amp-story-education/0.1/amp-story-education.js
@@ -26,7 +26,7 @@ import {Services} from '../../../src/services';
 import {createShadowRootWithStyle} from '../../amp-story/1.0/utils';
 import {dev} from '../../../src/log';
 import {dict} from '../../../src/utils/object';
-import {getLocalizationService} from '../../amp-story/1.0/amp-story';
+import {getLocalizationService} from '../../amp-story/1.0/amp-story-localization-service';
 import {htmlFor} from '../../../src/static-template';
 import {removeChildren} from '../../../src/dom';
 import {toggle} from '../../../src/style';
