+++ extensions/amp-story/1.0/simple-template.js
@@ -16,7 +16,7 @@
 import {LocalizedStringId} from '../../../src/localized-strings'; // eslint-disable-line no-unused-vars
 import {createElementWithAttributes} from '../../../src/dom';
 import {devAssert} from '../../../src/log';
-import {getLocalizationService} from './amp-story';
+import {getLocalizationService} from './amp-story-localization-service';
 import {hasOwn} from '../../../src/utils/object';
 import {isArray} from '../../../src/types';
